# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orchis Foundry API",
    "version": "0.7.0",
    "description": "Self-hosted code platform: git hosting, pull requests, per-user BYO-model supply-chain scanning, and LLM-friendly context endpoints. Authenticate with a personal access token: `Authorization: Bearer orc_pat_<...>`."
  },
  "servers": [{ "url": "/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "orc_pat" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/repos/{org}/{name}/proposals": {
      "get": { "summary": "List change proposals", "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "accepted", "rejected"] } }], "responses": { "200": { "description": "[Proposal]" } } },
      "post": { "summary": "Propose a change (human-in-the-loop gate; needs write access)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "summary": { "type": "string" }, "patch": { "type": "string" }, "changes": { "type": "array", "items": { "type": "object" } }, "baseSha": { "type": "string" }, "targetBranch": { "type": "string" } }, "required": ["title"] } } } }, "responses": { "200": { "description": "{id,status}" } } }
    },
    "/repos/{org}/{name}/proposals/{id}": { "get": { "summary": "Proposal detail", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Proposal" } } } },
...
  }
}